+++ CHANGELOG.md
@@ -6,6 +6,47 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.21.2] - 2025-06-13
+### Removed
+- Temporarily removed the Bionetta demo link.
+
+## [5.21.1] - 2025-06-11
+### Added
+- Link to a Bionetta demo.
+
+## [5.21.0] - 2025-06-05
+### Changed
+- Renamed ZK Reputation to ZK Graph;
+- Renamed zk-reputation.png image to zk-graph.png;
+
+## [5.20.0] - 2025-06-04
+### Added
+- ZK Likeness overview;
+- ZK Liveness overview;
+
+## [5.19.0] - 2025-06-04
+### Fixed
+- Fixed Seedless Recovery overview based on feedback;
+
+## [5.18.2] - 2025-06-02
+### Fixed
+- Noir is coming soon;
+
+## [5.18.0] - 2025-06-02
+### Added
+- Updated the Bionetta overview;
+- Updated the Bionetta training guide;
+
+## [5.17.0] - 2025-05-30
+### Added
+- Seedless Recovry overview;
+- "Solutions" section in the sidebar;
+
+## [5.16.1] - 2025-05-21
+### Fixed
+- Telegram links;
+
+## [5.16.0] - 2025-05-21
 ### Added
 - Bionetta training guide;
 
@@ -157,7 +198,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [5.0.0] - 2024-11-21
 ### Added
 - ZK Registry docs;
-- ZK Reputation docs;
+- ZK Graph docs;
 - ZK Passport docs;
 
 ### Changed
@@ -554,7 +595,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Broken logo link in the footer;
 
 
-[Unreleased]: https://github.com/rarimo/docs/compare/5.15.0...HEAD
+[Unreleased]: https://github.com/rarimo/docs/compare/5.21.2...HEAD
+[5.21.2]: https://github.com/rarimo/docs/compare/5.21.1...5.21.2
+[5.21.1]: https://github.com/rarimo/docs/compare/5.21.0...5.21.1
+[5.21.0]: https://github.com/rarimo/docs/compare/5.20.0...5.21.0
+[5.20.0]: https://github.com/rarimo/docs/compare/5.19.0...5.20.0
+[5.19.0]: https://github.com/rarimo/docs/compare/5.18.2...5.19.0
+[5.18.2]: https://github.com/rarimo/docs/compare/5.18.0...5.18.2
+[5.18.0]: https://github.com/rarimo/docs/compare/5.17.0...5.18.0
+[5.17.0]: https://github.com/rarimo/docs/compare/5.16.1...5.17.0
+[5.16.1]: https://github.com/rarimo/docs/compare/5.16.0...5.16.1
+[5.16.0]: https://github.com/rarimo/docs/compare/5.15.0...5.16.0
 [5.15.0]: https://github.com/rarimo/docs/compare/5.14.0...5.15.0
 [5.14.0]: https://github.com/rarimo/docs/compare/5.13.0...5.14.0
 [5.13.0]: https://github.com/rarimo/docs/compare/5.12.0...5.13.0